requirements.txt
absl-py==1.3.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.24.93
botocore==1.27.93
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
Django==4.1.2
django-appconf==1.0.5
django-environ==0.9.0
entrypoints==0.4
executing==1.1.1
fire==0.4.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.5
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
rsa==4.9
s3transfer==0.6.0
sahi==0.10.7
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.1
Shapely==1.8.5.post1
six==1.16.0
sklearn==0.0
sqlparse==0.4.3
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
yolov5==6.2.2
zipp==3.9.0
